Move tests to on folder
